+++ internal/codec/jp2k/jp2k.go
@@ -59,6 +59,7 @@ type Encoder struct {
 
 func (*Encoder) LevelHeader() []byte        { return nil }
 func (*Encoder) TIFFCompressionTag() uint16 { return tiff.CompressionJPEG2000 }
+func (*Encoder) TIFFPhotometric() uint16    { return codec.PhotometricRGB }
 func (*Encoder) Close() error               { return nil }
 
 // IsLossless reports whether this encoder produces byte-exact (reversible 5/3) output.